refactor(ui): split app shell bars into subcomponents
