+++ src/components/bottom-bar/BottomBarActions.tsx
@@ -1,0 +1,32 @@
+import { CogIcon } from 'lucide-react';
+import { Button } from '@/components/ui/button';
+import { cn } from '@/lib/utils';
+import { useViewControls } from '@/components/ViewControlsContext';
+
+export function BottomBarActions() {
+    const { isViewControlsOpen, toggleViewControls } = useViewControls();
+
+    return (
+        <div className='mx-auto flex min-h-14 max-w-screen-sm items-center justify-between gap-4 px-4 py-3 text-sm font-semibold'>
+            <div className='flex flex-1 items-center justify-start'>Chat</div>
+            <div className='flex flex-1 items-center justify-center'>
+                <Button
+                    size='sm'
+                    variant={isViewControlsOpen ? 'default' : 'outline'}
+                    type='button'
+                    onClick={toggleViewControls}
+                    className={cn(
+                        'gap-2 text-[11px] font-semibold uppercase tracking-wide',
+                        isViewControlsOpen ?
+                            'bg-blue-700 text-white hover:bg-blue-600'
+                        :   'focus-visible:outline focus-visible:outline-2 focus-visible:outline-dashed focus-visible:outline-blue-400 focus-visible:outline-offset-2'
+                    )}
+                >
+                    <CogIcon className='h-4 w-4' />
+                    View Controls
+                </Button>
+            </div>
+            <div className='flex flex-1 items-center justify-end'>Population</div>
+        </div>
+    );
+}
