+++ src/components/top-bar/TopBarPreferencesDialog.tsx
@@ -1,0 +1,80 @@
+import * as React from 'react';
+import { Switch } from '@/components/ui/switch';
+import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle } from '@/components/ui/dialog';
+import { Button } from '@/components/ui/button';
+import { Settings } from 'lucide-react';
+import { useAppDispatch, useAppSelector } from '@/store/hooks';
+import {
+    selectShowFirstNightOrder,
+    selectShowNightOrder,
+    selectShowOtherNightOrder,
+    setShowFirstNightOrder,
+    setShowNightOrder,
+    setShowOtherNightOrder
+} from '@/store/settings/settings-slice';
+
+export function TopBarPreferencesDialog() {
+    const dispatch = useAppDispatch();
+    const showNightOrder = useAppSelector(selectShowNightOrder);
+    const showFirstNightOrder = useAppSelector(selectShowFirstNightOrder);
+    const showOtherNightOrder = useAppSelector(selectShowOtherNightOrder);
+    const [isSettingsOpen, setIsSettingsOpen] = React.useState(false);
+
+    return (
+        <Dialog
+            open={isSettingsOpen}
+            onOpenChange={setIsSettingsOpen}
+        >
+            <Button
+                variant='ghost'
+                className='h-9 px-3 gap-2'
+                type='button'
+                onClick={() => setIsSettingsOpen(true)}
+            >
+                <Settings className='h-4 w-4' />
+                Settings (Preferences)
+            </Button>
+            <DialogContent className='sm:max-w-lg'>
+                <DialogHeader>
+                    <DialogTitle>Preferences</DialogTitle>
+                    <DialogDescription>Adjust game UI visibility settings.</DialogDescription>
+                </DialogHeader>
+                <div className='space-y-4'>
+                    <div className='border-b border-border/60 pb-3 text-xs font-semibold uppercase tracking-wide text-muted-foreground'>
+                        Night order badges
+                    </div>
+                    <div className='flex items-center justify-between gap-4'>
+                        <div>
+                            <div className='text-sm font-semibold'>Show night order badges</div>
+                            <div className='text-xs text-muted-foreground'>Toggle all night order indicators.</div>
+                        </div>
+                        <Switch
+                            checked={showNightOrder}
+                            onCheckedChange={(value) => dispatch(setShowNightOrder(value))}
+                        />
+                    </div>
+                    <div className='flex items-center justify-between gap-4'>
+                        <div>
+                            <div className='text-sm font-semibold'>Show first-night order</div>
+                            <div className='text-xs text-muted-foreground'>Show the blue first-night badges.</div>
+                        </div>
+                        <Switch
+                            checked={showFirstNightOrder}
+                            onCheckedChange={(value) => dispatch(setShowFirstNightOrder(value))}
+                        />
+                    </div>
+                    <div className='flex items-center justify-between gap-4'>
+                        <div>
+                            <div className='text-sm font-semibold'>Show other-night order</div>
+                            <div className='text-xs text-muted-foreground'>Show the red other-night badges.</div>
+                        </div>
+                        <Switch
+                            checked={showOtherNightOrder}
+                            onCheckedChange={(value) => dispatch(setShowOtherNightOrder(value))}
+                        />
+                    </div>
+                </div>
+            </DialogContent>
+        </Dialog>
+    );
+}
